pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Web task manager written in django"
authors = ["Igor Kvachenok <darestade@gmail.com>"]
readme = "README.md"
packages = [{include = "task_manager"}]

[tool.poetry.dependencies]
python = "^3.8.1"
python-dotenv = "^1.0.0"
django = "^4.2.1"
gunicorn = "^20.1.0"
dj-database-url = "^0.5.0"
psycopg2-binary = "^2.9.6"
whitenoise = "^6.4.0"
django-bootstrap4 = "^23.1"
factory-boy = "^3.2.1"
rollbar = "^0.16.3"

[tool.poetry.group.dev.dependencies]
flake8 = "^4.0.0"
flake8-django = "^1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
